requirements.txt
numpy>=1.23
pandas>=2.0
pyyaml>=6.0
scikit-learn>=1.3
torch>=2.1
tqdm>=4.66
